requirements.txt
# Django Logistics System Requirements
# Optimized version - removed unused packages

# Core Django
django==5.1.7

# Database
psycopg2-binary==2.9.10

# WebSocket support (using InMemoryChannelLayer)
channels==4.2.0

# Static files
whitenoise==6.9.0

# Production server
gunicorn==23.0.0

# Environment variables
python-dotenv==1.0.1

# REST API
djangorestframework==3.15.2